WIP: Less startup and much better docs
