+++ src/Chrome.ts
@@ -1,149 +1,271 @@
-import * as _ from 'lodash';
-import * as chromeLauncher from 'chrome-launcher';
-import * as CDP from 'chrome-remote-interface';
+import { EventEmitter } from 'events';
+import * as path from 'path';
+import * as fs from 'fs';
 import * as debug from 'debug';
-
-import { ChromeTab, events } from './ChromeTab';
+import * as chrome from './util/chrome';
 
 const log = debug('navalia:chrome');
 
-export interface chromeOptions {
-  [propName: string]: boolean | undefined;
+export type triggerEvents =
+  'click' |
+  'mousedown' |
+  'mouseup' |
+  'mouseover' |
+  'touchstart' |
+  'touchend' |
+  'focus' |
+  'touchcancel' |
+  'touchmove' |
+  'change' |
+  'blur' |
+  'select'
+;
+
+export interface options {
+  flags: chrome.flags,
+  cdp: chrome.cdp,
 }
 
-export interface customOptions {
-  maxActiveTabs?: number;
+export interface navigateOpts {
+  onload?: boolean
 }
 
-export class Chrome {
-  private chrome: any;
-  private host: any;
-  private isExpired: boolean;
-  private activeTabs: number;
-  private maxActiveTabs: number;
-  private browserHasStarted: boolean;
-  private browserStartingPromise: Promise<any> | boolean;
-  private kill: Function;
+export const events = {
+  done: 'done',
+};
 
-  public chromeBootOptions: chromeOptions;
-  public jobsComplete: number;
-  public port: number;
+export class Chrome extends EventEmitter {
+  private cdp: chrome.cdp;
+  private kill: () => Promise<{}>;
+  private flags: chrome.flags;
 
-  constructor(chromeBootOptions: chromeOptions = {}, customOptions: customOptions = {}) {
-    this.isExpired = false;
-    this.browserStartingPromise = false;
-    this.browserHasStarted = false;
-    this.jobsComplete = 0;
-    this.activeTabs = 0;
-    this.maxActiveTabs = customOptions.maxActiveTabs || -1;
-    this.chromeBootOptions ={
-      headless: true,
-      disableGpu: true,
-      hideScrollbars: true,
-      ...chromeBootOptions,
-    };
+  constructor(options) {
+    super();
+    
+    if (options.cdp) {
+      this.cdp = options.cdp;
+    }
+
+    if (options.flags) {
+      this.flags = options.flags;
+    }
   }
 
-  private async getNewTab() {
-    const { browserContextId } = await this.host.Target.createBrowserContext();
+  private async getSelectorId(selector: string): Promise<number | null> {
+    const document = await this.cdp.DOM.getDocument();
 
-    log(`creating new tab at ${browserContextId}`);
-
-    const { targetId } = await this.host.Target.createTarget({
-      url: 'about:blank',
-      browserContextId
+    const { nodeId } = await this.cdp.DOM.querySelector({
+      nodeId: document.root.nodeId,
+      selector,
     });
 
-    // connct to the new context
-    const newTab = await CDP({ tab: `ws://localhost:${this.port}/devtools/page/${targetId}` });
-
-    // Enable all the domains
-    await Promise.all([
-      newTab.Page.enable(),
-      newTab.Runtime.enable(),
-      newTab.Network.enable(),
-      newTab.DOM.enable(),
-      newTab.CSS.enable(),
-    ]);
-
-    const tab = new ChromeTab(newTab, targetId);
-
-    // Fire an event when we're closed so Navalia
-    // can do cleanup
-    tab.on(events.done, this.onTabClose.bind(this));
-
-    this.activeTabs++;
-
-    return tab;
+    return nodeId;
   }
 
-  private async bootChrome() {
-    const chromeFlags = _.chain(this.chromeBootOptions)
-      .pickBy((value) => value)
-      .map((_value, key) => `--${_.kebabCase(key)}`)
-      .value();
-
-    log(`launching host with args ${chromeFlags.join(' ')}`);
-
-    // Boot Chrome
-    const browser = await chromeLauncher.launch({ chromeFlags });
-    const cdp = await CDP({ target: `ws://localhost:${browser.port}/devtools/browser` });
-
-    log(`launched host on port ${browser.port}`);
+  private async boot(): Promise<void> {
+    if (this.cdp) {
+      return;
+    }
+    log(`starting chrome`);
+    const { browser, cdp } = await chrome.launch(this.flags);
 
     this.kill = browser.kill;
-    this.host = cdp;
-    this.chrome = cdp;
-    this.port = browser.port;
-    this.browserHasStarted = true;
-    this.browserStartingPromise = false;
+    this.cdp = cdp;
+
+    return;
   }
 
-  public async start(): Promise<ChromeTab> {
-    // If browser has already started, return a fresh instance
-    if (this.browserHasStarted) {
-      return this.getNewTab();
+  private async trigger(eventName: triggerEvents, selector: string): Promise<any> {
+    let eventClass = '';
+
+    switch (eventName) {
+      case 'click':
+      case 'mousedown':
+      case 'mouseup':
+      case 'mouseover':
+        eventClass = 'MouseEvents';
+        break;
+
+      case 'touchstart':
+      case 'touchend':
+      case 'touchcancel':
+      case 'touchmove':
+        eventClass = 'TouchEvents';
+        break;
+
+      case 'focus':
+      case 'change':
+      case 'blur':
+      case 'select':
+        eventClass = 'HTMLEvents';
+        break;
+
+      default:
+        throw `chrome#trigger: Couldn't handle event ${eventName} on selector ${selector}.`;
     }
 
-    // If the browser is still starting, wait for it's completion
-    // then return a new tab
-    if (this.browserStartingPromise) {
-      await this.browserStartingPromise;
-      return this.getNewTab();
+    const expression = function(selector, eventName, eventClass) {
+      const node = document.querySelector(selector);
+      const doc = node && node.ownerDocument ? node.ownerDocument : node;
+      const e = doc && doc.createEvent(eventClass);
+      e.initEvent(eventName, true);
+      e.synthetic = true;
+      node.dispatchEvent(e, true);
+    };
+
+    return this.evaluate(expression, selector, eventName, eventClass);
+  }
+
+  public async navigate(url: string, opts: navigateOpts = { onload: true }): Promise<void> {
+    await this.boot();
+
+    log(`navigating to ${url}`);
+
+    await this.cdp.Page.navigate({ url });
+
+    if (opts.onload) {
+      await this.cdp.Page.loadEventFired();
+      return;
     }
 
-    // Browser hasn't started, so boot it and cache the promise
-    // so this method can be idempotent to consumers
-    this.browserStartingPromise = this.bootChrome();
-
-    await this.browserStartingPromise;
-
-    return this.getNewTab();
+    return;
   }
 
-  public onTabClose(targetId: string): void {
-    this.chrome.Target.closeTarget({ targetId });
-    log(`tab ${targetId} closed`);
-    this.activeTabs--;
+  public async evaluate(expression: Function, ...args): Promise<any> {
+    const stringifiedArgs = args.map((arg) => JSON.stringify(arg)).join(',');
+    const script = `(${expression.toString()})(${stringifiedArgs})`;
+    
+    log(`executing script: ${script}`);
+    
+    return this.cdp.Runtime.evaluate({ expression: script, returnValue: true });
   }
 
-  public async quit(): Promise<void> {
-    log(`killing instance`);
-    this.activeTabs = 0;
-    await this.chrome.close();
-    return this.kill();
+  public async screenshot(filePath: string): Promise<any> {
+    if (!path.isAbsolute(filePath)) {
+      throw new Error(`Filepath is not absolute: ${filePath}`);
+    }
+    log(`capturing screenshot ${filePath}`);
+
+    const base64Image = await this.cdp.Page.captureScreenshot();
+    const buffer = new Buffer(base64Image.data, 'base64');
+
+    return fs.writeFileSync(filePath, buffer, { encoding: 'base64' });
   }
 
-  public setExpired(): void {
-    log(`instance has been marked expired`);
-    this.isExpired = true;
+  public async pdf(filePath: string): Promise<any> {
+    if (!path.isAbsolute(filePath)) {
+      throw new Error(`Filepath is not absolute: ${filePath}`);
+    }
+    log(`capturing PDF ${filePath}`);
+
+    const base64Image = await this.cdp.Page.printToPDF();
+    const buffer = new Buffer(base64Image.data, 'base64');
+
+    return fs.writeFileSync(filePath, buffer, { encoding: 'base64' });
   }
 
-  public getIsBusy(): boolean {
-    return this.maxActiveTabs === this.activeTabs;
+  public async size(width:number, height:number): Promise<any> {
+    log(`setting window size ${width}x${height}`);
+
+    return this.cdp.Emulation.setVisibleSize({ width: width, height: height });
   }
 
-  public getIsExpired(): boolean {
-    return this.isExpired;
+  public async exists(selector: string): Promise<boolean> {
+    log(`checking if '${selector}' exists`);
+    
+    return !!await this.getSelectorId(selector);
+  }
+
+  public async html(selector: string): Promise<string | null> {
+    log(`getting '${selector}' HTML`);
+
+    const nodeId = await this.getSelectorId(selector);
+
+    if (!nodeId) {
+      return null;
+    }
+
+    const { outerHTML } = await this.cdp.DOM.getOuterHTML({ nodeId });
+
+    return outerHTML;
+  }
+
+  public async click(selector: string): Promise<void> {
+    log(`clicking '${selector}'`);
+    return this.trigger('click', selector);
+  }
+
+  public async focus(selector: string): Promise<void> {
+    log(`focusing '${selector}'`);
+    return this.trigger('focus', selector);
+  }
+
+  public async type(selector:string, value:string): Promise<void> {
+    log(`typing'${value}' into '${selector}'`);
+    return this.evaluate((selector, value) => {
+      var element = document.querySelector(selector);
+      if (element) {
+        element.value = value;
+      }
+    }, selector, value);
+  }
+
+  public async check(selector:string): Promise<void> {
+    log(`checking checkbox '${selector}'`);
+    return this.evaluate((selector) => {
+      var element = document.querySelector(selector);
+      if (element) {
+        element.checked = true;
+      }
+    }, selector);
+  }
+
+  public async uncheck(selector:string): Promise<void> {
+    log(`un-checking checkbox '${selector}'`);
+    return this.evaluate((selector) => {
+      var element = document.querySelector(selector);
+      if (element) {
+        element.checked = false;
+      }
+    }, selector);
+  }
+
+  public async select(selector: string, option:string): Promise<void> {
+    log(`selecting option '${option}' in '${selector}'`);
+    return this.evaluate((selector) => {
+      var element = document.querySelector(selector);
+      if (element) {
+        element.value = option;
+      }
+    }, selector);
+  }
+
+  public async visible(selector: string): Promise<boolean> {
+    log(`seeing if '${selector}' is visible`);
+    return this.evaluate((selector) => {
+      var element = document.querySelector(selector);
+      if (element) {
+        return (element.offsetWidth > 0 && element.offsetHeight > 0);
+      }
+      else return false;
+    }, selector);
+  }
+
+  public async wait(time: number): Promise<object> {
+    log(`waiting ${time} ms`);
+    return new Promise((resolve) => { 
+      setTimeout(() => resolve(), time);
+    });
+  }
+
+  public done(): void {
+    log(`finished`);
+
+    if (this.kill) {
+      log(`closing chrome`);
+      this.kill();
+    }
+
+    this.emit(events.done);
   }
 }
